refactoring/relocation
